requirements.txt
Django==2.1.7
djangorestframework_jwt==1.11.0
PyJWT==1.7.1
djangorestframework==3.10.2
celery==4.3.0